requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
scikit-learn
joblib
numpy
python-multipart